services/service4/receiver/requirements.txt
mysql-connector-python==8.0.23
pika==1.2.0